+++ codegen/__out__/pc/src/views/wxwork/wxw_usr/Api.ts
@@ -198,9 +198,12 @@ export async function updateById(
  * 根据 id 查找企微用户
  */
 export async function findById(
-  id: WxwUsrId,
+  id?: WxwUsrId,
   opt?: GqlOpt,
-) {
+): Promise<WxwUsrModel | undefined> {
+  if (!id) {
+    return;
+  }
   const data: {
     findByIdWxwUsr?: WxwUsrModel;
   } = await query({
@@ -221,12 +224,52 @@ export async function findById(
 }
 
 /**
+ * 根据 ids 查找企微用户
+ */
+export async function findByIds(
+  ids: WxwUsrId[],
+  opt?: GqlOpt,
+): Promise<WxwUsrModel[]> {
+  if (ids.length === 0) {
+    return [ ];
+  }
+  opt = opt || { };
+  opt.showErrMsg = false;
+  let models: WxwUsrModel[] = [ ];
+  try {
+    const data: {
+      findByIdsWxwUsr: WxwUsrModel[];
+    } = await query({
+      query: `
+        query($ids: [WxwUsrId!]!) {
+          findByIdsWxwUsr(ids: $ids) {
+            ${ wxwUsrQueryField }
+          }
+        }
+      `,
+      variables: {
+        ids,
+      },
+    }, opt);
+    models = data.findByIdsWxwUsr;
+  } catch (_err) { /* empty */ }
+  for (let i = 0; i < models.length; i++) {
+    const model = models[i];
+    await setLblById(model);
+  }
+  return models;
+}
+
+/**
  * 根据 ids 删除企微用户
  */
 export async function deleteByIds(
   ids: WxwUsrId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     deleteByIdsWxwUsr: Mutation["deleteByIdsWxwUsr"];
   } = await mutation({
@@ -249,7 +292,10 @@ export async function deleteByIds(
 export async function revertByIds(
   ids: WxwUsrId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     revertByIdsWxwUsr: Mutation["revertByIdsWxwUsr"];
   } = await mutation({
@@ -272,7 +318,10 @@ export async function revertByIds(
 export async function forceDeleteByIds(
   ids: WxwUsrId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     forceDeleteByIdsWxwUsr: Mutation["forceDeleteByIdsWxwUsr"];
   } = await mutation({
